ptr size compensation when ptrs are pushed to the local script stack
there was a problem where ptrs that are pushed onto the local stack that were still thought of taking up only 4 bytes on the stack, so future script instructions may overwrite the upper 4 bytes of the ptr written on the stack.

solution isn't elegant and its probably not in the best place it should be right now, requires little more research. however, with this fix, we can now load admin into gameplay. still crashing at the end of the intro cutscene.